Remove dead class/cleanup. Rename AtlasClientFactory to AtlasDialogueServiceCreator. Plumb HostEventsSink (remains to be seen if we can actually pull this through).
